pyproject.toml
[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[project]
name = "ibis-crew-ai"
version = "0.1.0"
description = "Data manipulation using ibis with AI"
readme = "README.md"
dependencies = [
    "langchain-google-vertexai~=2.0.7",
    "langchain~=0.3.14",
    "langchain-community~=0.3.17",
    "langchain-openai~=0.3.5",
    "langgraph~=0.2.63",
    "crewai~=0.102.0",
    "langchain-core~=0.3.9",
    "opentelemetry-exporter-gcp-trace~=1.9.0",
    "traceloop-sdk~=0.38.7",
    "google-cloud-logging~=3.11.4",
    "google-cloud-aiplatform[evaluation]~=1.87.0",
    "fastapi~=0.115.8",
    "uvicorn~=0.34.0",
    "loguru>=0.7.3",
    "dynaconf>=3.2.10",
    "quartodoc>=0.9.1",
]

requires-python = ">=3.10,<3.14"


[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.23.8",
    "nest-asyncio>=1.6.0",
    "locust>=2.34.1",
    "pytest-cov>=6.1.1",
]

[project.optional-dependencies]
streamlit = [
    "streamlit~=1.42.0",
    "streamlit-extras~=0.4.3",
    "extra-streamlit-components~=0.1.71",
    "streamlit-feedback~=0.1.3",
]
jupyter = [
    "jupyter~=1.0.0",
]
lint = [
    "ruff>=0.4.6",
    "mypy~=1.15.0",
    "codespell~=2.2.0",
    "types-pyyaml~=6.0.12.20240917",
    "types-requests~=2.32.0.20240914",
]

[tool.ruff]
src = [ "src/ibis_crew_ai", "tests", "conf",]
line-length = 120
exclude = [ ".git", ".ruff_cache", ".venv", "__pypackages__", "__init__.py",]

[tool.autopep8]
max_line_length = 120

[tool.ruff.lint]
select = [ "ALL",]
ignore = [ ]

[tool.ruff.lint.extend-per-file-ignores]
"tests/**/*.py" = ["S101"] # assert used in tests

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.flake8-quotes]
inline-quotes = "single"

[tool.hatch.build.targets.wheel]
packages = ["src/ibis_crew_ai", "src/conf", "src/frontend"]

